Fix application failure
The Kafdrop chart was unable to successfully create a working service
due to a number of issues.

Setting the root filesystem to read only prevented the application from
writing to /tmp.

Running the application as a non root user caused the web interface to
hang when rendering the main view.

The ingress was incorrectly configured and caused Kubernetes to output
a schema failure error.